1st edit

+++ Lesson_01/triangle.rb
@@ -1,6 +1,6 @@
 puts "Введите длину основания треугольника"
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 puts "Введите высоту треугольника"
-h = gets.chomp.to_i
-sguare = a * h / 2
+h = gets.chomp.to_f
+sguare = a * h / 2.0
 puts "Площадь треугольника составляет #{sguare}"
